Let every walk driver be told which pool to walk (#68)
simulate_bloch, _simulate_bloch_mt and simulate_cpmg all seeded
geometry.init_positions(n, key) with no way to override it, and Mesh.init_positions
defaults to intra=True -- inside the surface. On a sphere that guess is right, which
is why nobody noticed; on a fibre bundle's extra-axonal pool, whose geometry is the
OUTER surface, it silently walks the intra pool instead. That is #67, measured there
at 97% of "extra" walkers ending inside an axon.

The seeding now goes through geometries.initial_positions, which replaced four inline
copies -- two of which validated the shape and two of which did not -- so "which pool
did this run walk?" has one answer and one validation instead of one per driver.

Three defects fell out of writing the test for it:

simulate_cpmg raised NameError: sub_steps on EVERY call. Mine, from eb7477b: I added
sub_steps=sub_steps to its make_step_fn call without adding the parameter, and
tests/test_cpmg_multiecho.py has been failing since. The parameter is real now.

Mesh's three bounce loops pinned their scan carry to int32 while argmin returns int64
once anything in the process enables x64 -- which bloch._simulate_bloch_mt does,
globally and permanently, for its float64 magnetisation. Any mesh walk after an MT
Bloch call in the same process died on a carry dtype mismatch, so test outcomes were
order-dependent. idx is now cast explicitly; a facet index needs 31 bits for 2e9
triangles, so int32 costs nothing. The config leak itself is #70.

Plain simulate_bloch accepts sub_steps and ignores it, so it cannot confine a mesh at
all: 0.05052 on a 2 um mesh sphere where core.simulate gives 0.96305 and the analytic
Sphere 0.96442, against a free limit of 0.01867. Not fixed here -- #69 -- but it is
why the new test uses an analytic Sphere and the start distribution rather than a mesh
and pool contrast. A pool test on a mesh would have measured that bug instead of this
one. Mesh pool contrast is covered on the trajectory driver, which handles it.

Verified: test_seed_pool_plumbing 4, test_cpmg_multiecho 2, test_mt_mesh_binding 2,
test_bloch_forward + test_bloch_permeation + test_mt_binding + test_mt_oracle 25, and
the six mesh suites 45.

+++ dmipy_sim/mesh.py
@@ -786,7 +786,12 @@ class Mesh(Geometry):
             # bounce instead of one per step. Without the sub-step cap it would NOT be sound.
             ts, u, v = self._mt(r0, dh, tri, valid)
             vm = (ts > self._hit_floor(i > 0)) & (ts < rem); ts = jnp.where(vm, ts, jnp.inf)
-            idx = jnp.argmin(ts); d = ts[idx]; hit = d < jnp.inf
+            # int32 explicitly: argmin returns int64 once anything in the process enables x64
+            # (`bloch._simulate_bloch_mt` flips that global toggle for its float64 magnetisation
+            # and never restores it), and the scan carry below is pinned int32 -- a mismatch JAX
+            # rejects at trace time. Any mesh walk after an MT Bloch call in the same process hit
+            # it. The facet index needs 31 bits for 2e9 triangles, so int32 costs nothing.
+            idx = jnp.argmin(ts).astype(jnp.int32); d = ts[idx]; hit = d < jnp.inf
             d_ref, n, nf = self._bounce(vnf, nrmf, u, v, idx, dh)
             r_hit = r0 + d * dh
             # Nudge along the GEOMETRIC normal: it is perpendicular to the triangle just hit, so it
@@ -875,7 +880,12 @@ class Mesh(Geometry):
                 same = jnp.arange(ts.shape[0]) == last
                 vm = (ts > 0.0) & (ts < rem) & jnp.logical_not(same & (ts < self._REST_THR))
             ts = jnp.where(vm, ts, jnp.inf)
-            idx = jnp.argmin(ts); d = ts[idx]; hit = d < jnp.inf
+            # int32 explicitly: argmin returns int64 once anything in the process enables x64
+            # (`bloch._simulate_bloch_mt` flips that global toggle for its float64 magnetisation
+            # and never restores it), and the scan carry below is pinned int32 -- a mismatch JAX
+            # rejects at trace time. Any mesh walk after an MT Bloch call in the same process hit
+            # it. The facet index needs 31 bits for 2e9 triangles, so int32 costs nothing.
+            idx = jnp.argmin(ts).astype(jnp.int32); d = ts[idx]; hit = d < jnp.inf
             d_ref, n, nf = self._bounce(vnf, nrmf, u, v, idx, dh)
             # side of the collision: dh·(outward face normal) > 0 -> spin leaving a
             # cell (intra side), < 0 -> entering (extra side).
@@ -952,7 +962,12 @@ class Mesh(Geometry):
             # tests. That was wrong: they fail identically at 6310a89, before any of this work, and they
             # exercise the analytic `Sphere` geometry rather than a mesh, so nothing here can reach them.)
             vm = (ts > self._EPS) & (ts < rem); ts = jnp.where(vm, ts, jnp.inf)
-            idx = jnp.argmin(ts); d = ts[idx]; hit = d < jnp.inf
+            # int32 explicitly: argmin returns int64 once anything in the process enables x64
+            # (`bloch._simulate_bloch_mt` flips that global toggle for its float64 magnetisation
+            # and never restores it), and the scan carry below is pinned int32 -- a mismatch JAX
+            # rejects at trace time. Any mesh walk after an MT Bloch call in the same process hit
+            # it. The facet index needs 31 bits for 2e9 triangles, so int32 costs nothing.
+            idx = jnp.argmin(ts).astype(jnp.int32); d = ts[idx]; hit = d < jnp.inf
             d_ref, n, nf = self._bounce(vnf, nrmf, u, v, idx, dh)
             # crossing direction: dh·(outward normal) > 0 -> leaving a cell
             # (intra->extra), < 0 -> entering (extra->intra).
